Created method PUT and refactored methods Create and class MessageResponseDTO
