+++ src/main/java/one/digitalinnovation/personapi/dto/response/MessageResponseDTO.java
@@ -1,10 +1,10 @@
-package one.digitalinnovation.personapi.dto;
+package one.digitalinnovation.personapi.dto.response;
 
 import lombok.Builder;
 import lombok.Data;
 
 @Data
 @Builder
-public class MessagePersonDTO {
+public class MessageResponseDTO {
     private String message;
 }
