+++ src/main/java/one/digitalinnovation/personapi/service/PersonService.java
@@ -1,6 +1,6 @@
 package one.digitalinnovation.personapi.service;
 
-import one.digitalinnovation.personapi.dto.MessagePersonDTO;
+import one.digitalinnovation.personapi.dto.response.MessageResponseDTO;
 import one.digitalinnovation.personapi.dto.request.PersonDTO;
 import one.digitalinnovation.personapi.entity.Person;
 import one.digitalinnovation.personapi.exception.PersonNotFoundException;
@@ -11,7 +11,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -26,15 +25,12 @@ public class PersonService {
         this.personRepository = personRepository;
     }
 
-    public MessagePersonDTO createPerson(@RequestBody PersonDTO personDTO) {
-        Person personSave = personMapper.toModel(personDTO);
+    public MessageResponseDTO createPerson(@RequestBody PersonDTO personDTO) {
+        Person personToSave = personMapper.toModel(personDTO);
 
-        Person savedPerson = personRepository.save(personSave);
+        Person savedPerson = personRepository.save(personToSave);
 
-        return MessagePersonDTO
-                .builder()
-                .message("Created person with Id " + savedPerson.getId())
-                .build();
+        return createMessageResponse(savedPerson.getId(), "Created person with Id ");
     }
 
     public List<PersonDTO> listAll() {
@@ -56,8 +52,27 @@ public class PersonService {
         personRepository.deleteById(id);
     }
 
+    public MessageResponseDTO update(Long id, PersonDTO personDTO) throws PersonNotFoundException {
+
+        verifyIfExists(id);
+
+        Person personToUpdate = personMapper.toModel(personDTO);
+
+        Person updatedPerson = personRepository.save(personToUpdate);
+
+        return createMessageResponse(updatedPerson.getId(), "Updated person with Id ");
+    }
+
     private Person verifyIfExists(Long id) throws PersonNotFoundException {
         return personRepository.findById(id)
                 .orElseThrow(() -> new PersonNotFoundException(id));
+    }
+
+
+    private MessageResponseDTO createMessageResponse(Long id, String message) {
+        return MessageResponseDTO
+                .builder()
+                .message(message + id)
+                .build();
     }
 }
